Split Item related classes out of ItemManager
