+++ GameServerCore/Domain/IItem.cs
@@ -3,6 +3,7 @@
     public interface IItem
     {
         byte StackSize { get; }
+        // todo: Unused atm, is this needed?
         int TotalPrice { get; }
         IItemType ItemType { get; }
     }
