+++ GameServerLib/Items/ItemType.cs
@@ -1,59 +1,9 @@
-﻿using System.Collections.Generic;
-using System.Linq;
 using GameServerCore.Domain;
+using LeagueSandbox.GameServer.Content;
 using LeagueSandbox.GameServer.GameObjects.Stats;
-using LeagueSandbox.GameServer.Items;
 
-namespace LeagueSandbox.GameServer.Content
+namespace LeagueSandbox.GameServer.Items
 {
-    public class ItemManager
-    {
-        private Dictionary<int, ItemType> _itemTypes;
-
-        public ItemManager()
-        {
-            _itemTypes = new Dictionary<int, ItemType>();
-        }
-
-        public ItemType GetItemType(int itemId)
-        {
-            return _itemTypes[itemId];
-        }
-
-        public ItemType SafeGetItemType(int itemId, ItemType defaultValue)
-        {
-            if (!_itemTypes.ContainsKey(itemId))
-            {
-                return defaultValue;
-            }
-
-            return _itemTypes[itemId];
-        }
-
-        public ItemType SafeGetItemType(int itemId)
-        {
-            return SafeGetItemType(itemId, null);
-        }
-
-        public void ResetItems()
-        {
-            _itemTypes.Clear();
-        }
-
-        public void LoadItems(string contentPath)
-        {
-            var itemContentCollection = ItemContentCollection.LoadItemsFrom(
-                $"{contentPath}/LeagueSandbox-Default/Items"
-            );
-
-            foreach (var entry in itemContentCollection)
-            {
-                var itemType = ItemType.Load(this, entry.Value);
-                _itemTypes.Add(entry.Key, itemType);
-            }
-        }
-    }
-
     public class ItemType : StatsModifier, IItemType
     {
         //private ItemManager _owner;
@@ -185,117 +135,4 @@ namespace LeagueSandbox.GameServer.Content
         }
     }
 
-    public class ItemRecipe
-    {
-        private ItemType _owner;
-        private ItemType[] _items;
-        private int _totalPrice;
-        private ItemManager _itemManager;
-
-        public int TotalPrice
-        {
-            get
-            {
-                if (_totalPrice <= -1)
-                {
-                    FindPrice();
-                }
-
-                return _totalPrice;
-            }
-        }
-
-        private ItemRecipe(ItemType owner, ItemManager manager)
-        {
-            _owner = owner;
-            _totalPrice = -1;
-            _itemManager = manager;
-        }
-
-        public List<ItemType> GetItems()
-        {
-            if (_items == null)
-            {
-                FindRecipeItems(_itemManager);
-            }
-
-            return _items.ToList();
-        }
-
-        private void FindRecipeItems(ItemManager itemManager)
-        {
-            // TODO: Figure out how to refactor this.
-            _items = new[]
-            {
-                itemManager.SafeGetItemType(_owner.RecipeItem1),
-                itemManager.SafeGetItemType(_owner.RecipeItem2),
-                itemManager.SafeGetItemType(_owner.RecipeItem3),
-                itemManager.SafeGetItemType(_owner.RecipeItem4)
-            };
-        }
-
-        private void FindPrice()
-        {
-            _totalPrice = 0;
-            foreach (var item in GetItems())
-            {
-                if (item != null)
-                {
-                    _totalPrice += item.TotalPrice;
-                }
-            }
-
-            _totalPrice += _owner.Price;
-        }
-
-        public static ItemRecipe FromItemType(ItemType type, ItemManager manager)
-        {
-            return new ItemRecipe(type, manager);
-        }
-    }
-
-    public class Item : IItem
-    {
-        public byte StackSize { get; private set; }
-        public int TotalPrice { get; private set; }
-        public ItemType ItemType { get; private set; }
-
-        IItemType IItem.ItemType => ItemType;
-
-        private Inventory _owner;
-
-        private Item(Inventory owner, ItemType type)
-        {
-            _owner = owner;
-            ItemType = type;
-            StackSize = 1;
-        }
-
-        public bool IncrementStackSize()
-        {
-            if (StackSize >= ItemType.MaxStack)
-            {
-                return false;
-            }
-
-            StackSize++;
-            return true;
-        }
-
-        public bool DecrementStackSize()
-        {
-            if (StackSize < 1)
-            {
-                return false;
-            }
-
-            StackSize--;
-            return true;
-        }
-
-        public static Item CreateFromType(Inventory inventory, ItemType item)
-        {
-            return new Item(inventory, item);
-        }
-    }
 }
