+++ GameServerLib/Packets/PacketHandlers/HandleSpawn.cs
@@ -1,7 +1,6 @@
 ﻿using ENet;
 using GameServerCore.Enums;
 using GameServerCore.Packets.Enums;
-using LeagueSandbox.GameServer.Content;
 using LeagueSandbox.GameServer.GameObjects;
 using LeagueSandbox.GameServer.GameObjects.AttackableUnits;
 using LeagueSandbox.GameServer.GameObjects.AttackableUnits.AI;
@@ -10,6 +9,7 @@ using LeagueSandbox.GameServer.GameObjects.Missiles;
 using LeagueSandbox.GameServer.Logging;
 using LeagueSandbox.GameServer.Players;
 using log4net;
+using LeagueSandbox.GameServer.Items;
 
 namespace LeagueSandbox.GameServer.Packets.PacketHandlers
 {
