+++ GameServerLib/Program.cs
@@ -1,8 +1,6 @@
-using System;
-using System.Timers;
-using LeagueSandbox.GameServer.Content;
 using LeagueSandbox.GameServer.Logging;
 using log4net;
+using LeagueSandbox.GameServer.Items;
 
 namespace LeagueSandbox.GameServer
 {
