+++ GameServerLibTests/Tests/Items/TestInventoryManager.cs
@@ -12,7 +12,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestAddAndGetItem()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var inventoryManager = InventoryManager.CreateInventory();
 
@@ -50,7 +50,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestItemStacking()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
 
@@ -90,7 +90,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestSetExtraItem()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
 
@@ -115,7 +115,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestGetItemSlot()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
 
@@ -146,7 +146,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestRemoveItem()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
 
@@ -171,7 +171,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         public void TestSwapItems()
         {
             var itemManager = new ItemManager();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
 
@@ -222,7 +222,7 @@ namespace LeagueSandbox.GameServerTests.Tests.Items
         {
             var itemManager = new ItemManager();
             itemManager.ResetItems();
-            itemManager.LoadItems("../../../Content");
+            itemManager.LoadItems(ItemContentCollection.LoadItemsFrom("../../../Content"));
 
             var manager = InventoryManager.CreateInventory();
